The "Now" page is looking more like o2m
The requests are now done by the browser using JSONP. The changes to the server were backwards compatible.

As a first test, this seems a much nicer way of doing things. Using data-bind all the time is OK because a lot of the logic that was previously inline can now be put into the "ES6" code.

The next test will be to see how events can be called, triggering comments and deletions.