Comments can now be added.
Individual pieces of content can reload their own html but they still cannot check if they have new children or if a child has been deleted.
It would be nice to just add the content in in the right place as if it had been pulled from the server. Likewise for deleting comments, it would be nice to just remove them when the server returns a success.

However, niether content nor links can be deleted yet. This should be pretty simple.